pyproject.toml
[tool.poetry]
name = "inaturalistbot"
version = "0.1.0"
description = "Telegram Bot to search taxa using iNaturalist"
authors = ["leynier <leynier41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.0"
pyinaturalist = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
